Add Point struct
Add a Point struct as a convenient holder for (x,y) indices. The Point
internally uses i64s in order to handle maps that place the origin at
some point other than the corner. Currently, the map only supports
indices starting at (0,0) but that is expected to change in the future.